+++ src/lvmopstools/devices/nps.py
@@ -8,7 +8,7 @@
 
 from __future__ import annotations
 
-from typing import TypedDict
+from typing_extensions import TypedDict
 
 from sdsstools import GatheringTaskGroup